move "mein Unterricht" parsing to new file
