+++ app/lib/client/client_submodules/mein_unterricht.dart
@@ -1,0 +1,373 @@
+import 'package:dio/dio.dart';
+import 'package:flutter/cupertino.dart';
+import 'package:html/parser.dart';
+import 'package:sph_plan/client/client.dart';
+
+import '../../shared/apps.dart';
+import '../../shared/shared_functions.dart';
+
+
+class MeinUnterrichtParser {
+  late Dio dio;
+  late SPHclient client;
+
+  MeinUnterrichtParser(Dio dioClient, this.client) {
+    dio = dioClient;
+  }
+
+
+  Future<dynamic> getOverview() async {
+    if (!client.doesSupportFeature(SPHAppEnum.meinUnterricht)) {
+      return -8;
+    }
+
+    debugPrint("Get Mein Unterricht overview");
+
+    var result = {"aktuell": [], "anwesenheiten": [], "kursmappen": []};
+
+    final response =
+    await dio.get("https://start.schulportal.hessen.de/meinunterricht.php");
+    var encryptedHTML = client.cryptor.decryptEncodedTags(response.data);
+    var document = parse(encryptedHTML);
+
+    //Aktuelle Einträge
+        () {
+      var schoolClasses = document.querySelectorAll("tr.printable");
+      for (var schoolClass in schoolClasses) {
+        var teacher = schoolClass.querySelector(".teacher");
+
+        if (schoolClass.querySelector(".datum") != null) {
+          result["aktuell"]?.add({
+            "name": schoolClass
+                .querySelector(".name")
+                ?.text
+                .trim(),
+            "teacher": {
+              "short": teacher
+                  ?.getElementsByClassName(
+                  "btn btn-primary dropdown-toggle btn-xs")[0]
+                  .text.trim(),
+              "name": teacher
+                  ?.querySelector("ul>li>a>i.fa")
+                  ?.parent
+                  ?.text
+                  .trim()
+            },
+            "thema": {
+              "title": schoolClass
+                  .querySelector(".thema")
+                  ?.text
+                  .trim(),
+              "date": schoolClass
+                  .querySelector(".datum")
+                  ?.text
+                  .trim()
+            },
+            "data": {
+              "entry": schoolClass.attributes["data-entry"],
+              "book": schoolClass.attributes["data-entry"]
+            },
+            "_courseURL": schoolClass
+                .querySelector("td>h3>a")
+                ?.attributes["href"]
+          });
+        }
+
+        //sort by date
+        result["aktuell"]?.sort((a, b) {
+          var aDate = a["thema"]["date"];
+          var bDate = b["thema"]["date"];
+
+          var aDateTime = DateTime(
+              int.parse(aDate.split(".")[2]),
+              int.parse(aDate.split(".")[1]),
+              int.parse(aDate.split(".")[0])
+          );
+          var bDateTime = DateTime(
+              int.parse(bDate.split(".")[2]),
+              int.parse(bDate.split(".")[1]),
+              int.parse(bDate.split(".")[0])
+          );
+
+          return bDateTime.compareTo(aDateTime);
+        });
+      }
+    }();
+
+    //Anwesenheiten
+    var anwesendDOM = document.getElementById("anwesend");
+        () {
+      var thead = anwesendDOM?.querySelector("thead>tr");
+      var tbody = anwesendDOM?.querySelectorAll("tbody>tr");
+
+      var keys = [];
+      thead?.children.forEach((element) => keys.add(element.text.trim()));
+
+      tbody?.forEach((elem) {
+        var textElements = [];
+        for (var i = 0; i < elem.children.length; i++) {
+          var element = elem.children[i];
+          element.querySelector("div.hidden.hidden_encoded")?.innerHtml = "";
+
+          if (keys[i] != "Kurs") {
+            textElements
+                .add(element.text.trim());
+          } else {
+            textElements.add(element.text.trim());
+          }
+        }
+
+        var rowEntry = {};
+
+        for (int i = 0; i < keys.length; i++) {
+          var key = keys[i];
+          var value = textElements[i];
+
+          rowEntry[key] = value;
+        }
+
+        //get url of course
+        var hyperlinkToCourse = elem.getElementsByTagName("a")[0];
+        rowEntry["_courseURL"] = hyperlinkToCourse.attributes["href"];
+
+        result["anwesenheiten"]?.add(rowEntry);
+      });
+    }();
+
+    //Kursmappen
+    var kursmappenDOM = document.getElementById("mappen");
+        () {
+      var parsedMappen = [];
+
+      var mappen = kursmappenDOM?.getElementsByClassName("row")[0].children;
+
+      if (mappen != null) {
+        for (var mappe in mappen) {
+          parsedMappen.add({
+            "title": mappe.getElementsByTagName("h2")[0].text.trim(),
+            "teacher":
+            mappe.querySelector("div.btn-group>button")?.attributes["title"],
+            "_courseURL":
+            mappe.querySelector("a.btn.btn-primary")?.attributes["href"]
+          });
+        }
+        result["kursmappen"] = parsedMappen;
+      } else {
+        result["kursmappen"] = [];
+      }
+
+    }();
+
+    debugPrint("Successfully got Mein Unterricht.");
+    return result;
+  }
+
+  Future<dynamic> setHomework(String courseID, String courseEntry, bool status) async {
+    //returns the response of the http request. 1 means success.
+
+    final response = await dio.post(
+      "https://start.schulportal.hessen.de/meinunterricht.php",
+      data: {
+        "a": "sus_homeworkDone",
+        "entry": courseEntry,
+        "id": courseID,
+        "b": status ? "done" : "undone"
+      },
+      options: Options(
+        headers: {
+          "Content-Type":
+          "application/x-www-form-urlencoded; charset=UTF-8",
+          "X-Requested-With": "XMLHttpRequest", //this is important
+        },
+      ),
+    );
+
+
+    return response.data;
+  }
+
+  Future<dynamic> getCourseView(String url) async {
+    try {
+      var result = {
+        "historie": [],
+        "leistungen": [],
+        "leistungskontrollen": [],
+        "anwesenheiten": [],
+        "name": ["name"],
+      };
+
+      String courseID = url.split("id=")[1];
+
+      final response =
+      await dio.get("https://start.schulportal.hessen.de/$url");
+      var encryptedHTML = client.cryptor.decryptEncodedTags(response.data);
+      var document = parse(encryptedHTML);
+
+      //course name
+      var heading = document.getElementById("content")?.querySelector("h1");
+      heading?.children[0].innerHtml = "";
+      result["name"] = [
+        heading?.text.trim()
+      ];
+
+      //historie
+          () {
+        var historySection = document.getElementById("history");
+        var tableRows = historySection?.querySelectorAll("table>tbody>tr");
+
+        tableRows?.forEach((tableRow) {
+          tableRow.children[2]
+              .querySelector("div.hidden.hidden_encoded")
+              ?.innerHtml = "";
+
+          Map<String, String> markups = {};
+
+          // There is also a css selector :has() but it's not implemented yet.
+          final String? content = tableRow.children[1].querySelector("span.markup i.far.fa-comment-alt:first-child")?.parent?.text.trim();
+          if (content != null && content.startsWith(" ")) {
+            markups["content"] = content.substring(1);
+          } else if (content != null) {
+            markups["content"] = content;
+          }
+
+          final String? homework = tableRow.children[1].querySelector("span.homework + br + span.markup")?.text.trim();
+          bool homeworkDone = false;
+
+          if (homework != null) {
+            homeworkDone = tableRow.querySelectorAll("span.done.hidden").isEmpty;
+            if (homework.startsWith(" ")) {
+              markups["homework"] = homework.substring(1);
+            } else {
+              markups["homework"] = homework;
+            }
+          }
+
+          List files = [];
+          if (tableRow.children[1].querySelector("div.alert.alert-info") != null) {
+            String baseURL = "https://start.schulportal.hessen.de/";
+            baseURL += tableRow.children[1].querySelector("div.alert.alert-info>a")!.attributes["href"]!;
+            baseURL = baseURL.replaceAll("&b=zip", "");
+
+            for (var fileDiv in tableRow.getElementsByClassName("files")[0].children) {
+              String? filename = fileDiv.attributes["data-file"];
+              files.add({
+                "filename": filename,
+                "filesize": fileDiv.querySelector("a>small")?.text,
+                "url": "$baseURL&f=$filename",
+              });
+            }
+          }
+
+          List uploads = [];
+          final uploadGroups = tableRow.children[1].querySelectorAll("div.btn-group");
+          for (final uploadGroup in uploadGroups) {
+            final openUpload = uploadGroup.querySelector(".btn-warning");
+            final closedUpload = uploadGroup.querySelector(".btn-default");
+
+            const String baseURL = "https://start.schulportal.hessen.de/";
+
+            if (openUpload != null) {
+              uploads.add({
+                "name": openUpload.nodes[2].text?.trim(),
+                "status": "open",
+                "link": baseURL + uploadGroup.querySelector("ul.dropdown-menu li a")!.attributes["href"]!,
+                "uploaded": openUpload.querySelector("span.badge")?.text,
+                "date": openUpload.querySelector("small")?.text.replaceAll("\n", "").replaceAll("                                                                ", "").replaceAll("bis ", "").replaceAll("um", ""),
+              });
+            } else if (closedUpload != null) {
+              uploads.add({
+                "name": closedUpload.nodes[2].text?.trim(),
+                "status": "closed",
+                "link": baseURL + uploadGroup.querySelector("ul.dropdown-menu li a")!.attributes["href"]!,
+                "uploaded": closedUpload.querySelector("span.badge")?.text,
+                "date": null,
+              });
+            }
+          }
+
+          result["historie"]?.add({
+            "time": tableRow.children[0].text.trim().replaceAll("  ", "").replaceAll("\n", " ").replaceAll("  ", " "),
+            "title": tableRow.children[1].querySelector("big>b")?.text.trim(),
+            "markup": markups,
+            "entry-id": tableRow.attributes["data-entry"],
+            "course-id": courseID,
+            "homework-done": homeworkDone,
+            "presence": tableRow.children[2].text.trim(),
+            "files": files,
+            "uploads": uploads
+          });
+        });
+      }();
+
+      //anwesenheiten
+          () {
+        var presenceSection = document.getElementById("attendanceTable");
+        var tableRows = presenceSection?.querySelectorAll("table>tbody>tr");
+
+        tableRows?.forEach((row) {
+          var encodedElements = row.getElementsByClassName("hidden_encoded");
+          for (var e in encodedElements) {
+            e.innerHtml = "";
+          }
+
+          result["anwesenheiten"]?.add(
+              {"type": row.children[0].text.trim(), "count": row.children[1].text.trim()});
+        });
+      }();
+
+      //leistungen
+          () {
+        var marksSection = document.getElementById("marks");
+        var tableRows = marksSection?.querySelectorAll("table>tbody>tr");
+
+        tableRows?.forEach((row) {
+          var encodedElements = row.getElementsByClassName("hidden_encoded");
+          for (var e in encodedElements) {
+            e.innerHtml = "";
+          }
+
+          if (row.children.length == 3) {
+            result["leistungen"]?.add({
+              "Name":
+              row.children[0].text.trim(),
+              "Datum":
+              row.children[1].text.trim(),
+              "Note": row.children[2].text.trim()
+            });
+          }
+        });
+      }();
+
+      //leistungskontrollen
+          () {
+        var examSection = document.getElementById("klausuren");
+        var lists = examSection?.children;
+
+        lists?.forEach((element) {
+          String exams = "";
+
+          final elements = element.querySelectorAll("ul li");
+
+          for (var element in elements) { //todo better solution that splitting the string with "  ...  "
+            var exam = element.text.trim().split("                                                                                                    ");
+            exams += "${exam.first.trim()} ${exam.last != exam.first ? exam.last.trim() : ""}";
+            if (element != elements.last) {
+              exams += "\n";
+            }
+          }
+
+          result["leistungskontrollen"]?.add({
+            "title": element.querySelector("h1,h2,h3,h4,h5,h6")?.text.trim(),
+            "value": exams == "" ? "Keine Daten!" : exams
+          });
+        });
+      }();
+
+      return result;
+    } catch (e, stack) {
+      recordError(e, stack);
+      return -4;
+    }
+  }
+
+}
